+++ pkg/cli/cmd/radinit/init_test.go
@@ -20,7 +20,6 @@ import (
 	"github.com/project-radius/radius/pkg/cli/aws"
 	"github.com/project-radius/radius/pkg/cli/azure"
 	"github.com/project-radius/radius/pkg/cli/clients"
-	cmdcommon "github.com/project-radius/radius/pkg/cli/cmd/radinit/common"
 	"github.com/project-radius/radius/pkg/cli/cmd/validation"
 	"github.com/project-radius/radius/pkg/cli/connections"
 	cli_credential "github.com/project-radius/radius/pkg/cli/credential"
@@ -123,7 +122,7 @@ func Test_Validate(t *testing.T) {
 						Name: to.Ptr("cool-existing-env"),
 					},
 				})
-				initExistingEnvironmentSelection(mocks.Prompter, cmdcommon.SelectExistingEnvironmentCreateSentinel)
+				initExistingEnvironmentSelection(mocks.Prompter, SelectExistingEnvironmentCreateSentinel)
 
 				// Use default env name and namespace
 				initEnvNamePrompt(mocks.Prompter)
@@ -808,13 +807,13 @@ func initEnvNamePromptError(prompter *prompt.MockInterface) {
 
 func initNamespacePrompt(prompter *prompt.MockInterface) {
 	prompter.EXPECT().
-		GetTextInput(cmdcommon.EnterNamespacePrompt, gomock.Any()).
+		GetTextInput(EnterNamespacePrompt, gomock.Any()).
 		Return("default", nil).Times(1)
 }
 
 func initNamespacePromptError(prompter *prompt.MockInterface) {
 	prompter.EXPECT().
-		GetTextInput(cmdcommon.EnterNamespacePrompt, gomock.Any()).
+		GetTextInput(EnterNamespacePrompt, gomock.Any()).
 		Return("", errors.New("Unable to read namespace")).Times(1)
 }
 
@@ -856,7 +855,7 @@ func setExistingEnvironments(clientMock *clients.MockApplicationsManagementClien
 
 func initExistingEnvironmentSelection(prompter *prompt.MockInterface, choice string) {
 	prompter.EXPECT().
-		GetListInput(gomock.Any(), cmdcommon.SelectExistingEnvironmentPrompt).
+		GetListInput(gomock.Any(), SelectExistingEnvironmentPrompt).
 		Return(choice, nil).Times(1)
 }
